// lib/types/company.ts
export interface CompanyInfo {
  id: string
  name: string
  name_kana: string | null
  postal_code: string | null
  address: string | null
  phone: string | null
  fax: string | null
  email: string | null
  invoice_number: string | null
  bank_name: string | null
  bank_branch: string | null
  bank_account_type: string | null
  bank_account_no: string | null
  bank_account_name: string | null
  updated_at: string
}